requirements.txt
playwright==1.40.0
anthropic==0.8.1
chromadb==0.4.18
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
scikit-learn==1.3.2
streamlit==1.34.0
